requirements.txt
# =========================
# Core dependencies
# =========================
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# =========================
# Computer Vision
# =========================
opencv-python>=4.8.0
ultralytics>=8.0.0

# =========================
# API & Web
# =========================
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
python-multipart>=0.0.6

# =========================
# LLM Integration
# =========================
anthropic>=0.7.0
openai>=1.3.0  # Optional: if you want to test GPT-4V too

# =========================
# Utilities
# =========================
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
tabulate>=0.9.0

# =========================
# Testing & Quality
# =========================
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
